requirements.txt
SpeechRecognition
groq
pillow
python-dotenv
google-generativeai
edge-tts
faster-whisper
opencv-python
streamlit
